shadow acne

+++ raytrace/src/color.rs
@@ -9,8 +9,8 @@ pub fn write_color(pixel_color:&vec3::Color, sample_per_pixel: i32) {
     let g = (pixel_color.y * scale).sqrt();
     let b = (pixel_color.z * scale).sqrt();
     println!("{} {} {}",
-             (256.0 * clamp(r, 0.0, 0.999)) as u8,
-             (256.0 * clamp(g, 0.0, 0.999)) as u8,
-             (256.0 * clamp(b, 0.0, 0.999)) as u8
+             (256.0 * clamp(r, 0.0, 0.999)) as i32,
+             (256.0 * clamp(g, 0.0, 0.999)) as i32,
+             (256.0 * clamp(b, 0.0, 0.999)) as i32
              )
 }
